move methods back off the module file
